+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchEffect.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.home.overdue.search
 
+import java.time.LocalDate
 import java.util.UUID
 
 sealed class OverdueSearchEffect
@@ -9,6 +10,8 @@ object LoadOverdueSearchHistory : OverdueSearchEffect()
 data class ValidateOverdueSearchQuery(val searchQuery: String) : OverdueSearchEffect()
 
 data class AddQueryToOverdueSearchHistory(val searchQuery: String) : OverdueSearchEffect()
+
+data class SearchOverduePatients(val searchQuery: String, val since: LocalDate) : OverdueSearchEffect()
 
 sealed class OverdueSearchViewEffect : OverdueSearchEffect()