+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Open overdue search screen when search button is clicked in overdue screen
 - Search for overdue patients when search query is valid
 - Turn off translation for instant search allowed characters
+- Updated translations: `om-ET`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300