+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Updated translations: `om-ET`, `pa-IN`, `kn-IN`, `so-ET`, `ti-ET`, `mr-IN`
 - Show search results when overdue search results are loaded
 - Handle user interactions for overdue search results
+- Implement overdue search results adapter
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300